# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/player": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Create a new player",
                "description": "Creates a new player with a name",
                "operationId": "createPlayer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "playerName"
                                ],
                                "properties": {
                                    "playerName": {
                                        "description": "Player's name",
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player created successfully"
                                        },
                                        "player": {
                                            "type": "string",
                                            "example": "stringified player object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/player/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get a player by ID",
                "description": "Returns a player by ID",
                "operationId": "getPlayerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the player to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Player stringified object"
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Player"
                ],
                "summary": "Update a player",
                "description": "Aggiorna le informazioni di un giocatore esistente, come nome, livello, esperienza raccolta, vittorie e battaglie totali.",
                "operationId": "2bee51cc75979ae669cc1b18558f0986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del giocatore da aggiornare",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nuovo nome del giocatore",
                                        "type": "string",
                                        "example": "john doe"
                                    },
                                    "level": {
                                        "description": "Nuovo livello del giocatore",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "experienceCollected": {
                                        "description": "Esperienza raccolta dal giocatore",
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "nWonBattles": {
                                        "description": "Numero di battaglie vinte dal giocatore",
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "nBattles": {
                                        "description": "Numero totale di battaglie del giocatore",
                                        "type": "integer",
                                        "example": 75
                                    },
                                    "helmetId": {
                                        "description": "ID del casco equipaggiato dal giocatore",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "runeId": {
                                        "description": "ID della runa equipaggiata dal giocatore",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "weaponId": {
                                        "description": "ID dell'arma equipaggiata dal giocatore",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Giocatore aggiornato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player updated successfully"
                                        },
                                        "player": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 22
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "john doe"
                                                },
                                                "level": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "experienceCollected": {
                                                    "type": "integer",
                                                    "example": 1000
                                                },
                                                "nWonBattles": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "nBattles": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Giocatore non trovato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Delete a player by ID",
                "description": "Deletes a player by ID",
                "operationId": "deletePlayerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the player to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player deleted successfully"
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}